moved featureset and distribution package to clg package
