+++ clg/distribution/common.go
