+++ clg/distribution/distribution_object.go
