+++ clg/distribution/distribution_object_test.go
