+++ clg/feature-set/feature.go
@@ -3,7 +3,7 @@ package featureset
 import (
 	"sync"
 
-	"github.com/xh3b4sd/anna/distribution"
+	"github.com/xh3b4sd/anna/clg/distribution"
 	"github.com/xh3b4sd/anna/id"
 	"github.com/xh3b4sd/anna/spec"
 )
