+++ clg/feature.go
@@ -1,7 +1,7 @@
 package clg
 
 import (
-	"github.com/xh3b4sd/anna/feature-set"
+	"github.com/xh3b4sd/anna/clg/feature-set"
 )
 
 func (i *clgIndex) AddPositionFeature(args ...interface{}) ([]interface{}, error) {
